+++ lib/modules/i18n/i18n_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: i18n
-// Updated at: 2026-08-03 21:22:34
-// Commit: fix(i18n): prevent double tap duplicate action in i18n
+// Updated at: 2026-08-03 23:23:25
+// Commit: feat(i18n): add bi-directional event bus handler for i18n
 
 class i18nComponentState {
-  final String commitRef = '3326';
-  final String lastMessage = 'fix(i18n): prevent double tap duplicate action in i18n';
-  final DateTime timestamp = DateTime.parse('2026-08-03T21:22:34+05:00');
+  final String commitRef = '3339';
+  final String lastMessage = 'feat(i18n): add bi-directional event bus handler for i18n';
+  final DateTime timestamp = DateTime.parse('2026-08-03T23:23:25+05:00');
 }